requirements.txt
numpy
pyTorch
scipy
matplotlib
pandas
nltk
